Hoist GHA input parameters to the constructor of OctopusCliWrapper
