+++ src/octopus-cli-wrapper.ts
@@ -1,21 +1,27 @@
+import { setFailed } from '@actions/core'
+import { exec, ExecOptions } from '@actions/exec'
 import { InputParameters } from './input-parameters'
 
 export class OctopusCliWrapper {
+  inputParameters: InputParameters
   // environment variables at the time the wrapper was created
   env: { [key: string]: string } | NodeJS.ProcessEnv
   logInfo: (message: string) => void
   logWarn: (message: string) => void
 
   constructor(
+    parameters: InputParameters,
     env: { [key: string]: string } | NodeJS.ProcessEnv,
     logInfo: (message: string) => void,
     logWarn: (message: string) => void
   ) {
+    this.inputParameters = parameters
     this.env = env
     this.logInfo = logInfo
     this.logWarn = logWarn
   }
 
+  // When the Octopus CLI writes to stdout, we capture the text via this function
   stdline(line: string) {
     if (line.length === 0) {
       return
@@ -52,6 +58,8 @@ export class OctopusCliWrapper {
     }
   }
 
+  // Picks up a config value from GHA Input or environment, supports mapping
+  // of an obsolete env var to a newer one (e.g. OCTOPUS_CLI_SERVER vs OCTOPUS_HOST)
   pickupConfigurationValueExtended(
     inputParameter: string,
     inputObsoleteEnvKey: string,
@@ -74,6 +82,7 @@ export class OctopusCliWrapper {
     }
   }
 
+  // Picks up a config value from GHA Input or environment
   pickupConfigurationValue(inputParameter: string, inputNewEnvKey: string, valueHandler: (value: string) => void) {
     if (inputParameter.length > 0) {
       valueHandler(inputParameter)
@@ -85,20 +94,14 @@ export class OctopusCliWrapper {
     }
   }
 
-  // mapping table:
-  // apiKey: getEnv(['OCTOPUS_HOST', 'OCTOPUS_CLI_API_KEY']),
-  // proxy: getEnv(['OCTOPUS_PROXY']),
-  // proxyPassword: getEnv(['OCTOPUS_PROXY_PASSWORD']),
-  // proxyUsername: getEnv(['OCTOPUS_PROXY_USERNAME']),
-  // server: getEnv(['OCTOPUS_HOST', 'OCTOPUS_CLI_SERVER']),
-  // space: getEnv(['OCTOPUS_SPACE'])
-
-  // Converts incoming environment and inputParameters into an array of strings to pass to the Octopus CLI
-  generateLaunchConfig(parameters: InputParameters): CliLaunchConfiguration {
+  // Converts incoming environment and inputParameters into a set of commandline args + env vars to run the Octopus CLI
+  generateLaunchConfig(): CliLaunchConfiguration {
     // Note: this is specialised to only work for create-release, but feels like it wants to be more generic and reusable?
     // Given we have multiple github actions and each lives in its own repo, what's our strategy for sharing here?
     const launchArgs: string[] = ['create-release']
     const launchEnv: { [key: string]: string } = {}
+
+    const parameters = this.inputParameters
 
     this.pickupConfigurationValueExtended(
       parameters.apiKey,
@@ -154,6 +157,29 @@ export class OctopusCliWrapper {
     }
 
     return { args: launchArgs, env: launchEnv }
+  }
+
+  // NOT UNIT TESTABLE. This shells out to 'octo' and expects to be running in GHA
+  // This invokes the CLI to do the work
+  async createRelease(): Promise<void> {
+    this.logInfo('🔣 Parsing inputs...')
+    const cliLaunchConfiguration = this.generateLaunchConfig()
+
+    const options: ExecOptions = {
+      listeners: {
+        stdline: this.stdline
+      },
+      env: cliLaunchConfiguration.env,
+      silent: true
+    }
+
+    try {
+      await exec('octo', cliLaunchConfiguration.args, options)
+    } catch (e: unknown) {
+      if (e instanceof Error) {
+        setFailed(e)
+      }
+    }
   }
 }
